requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
pip==23.1.2
setuptools==67.8.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
wheel==0.40.0
django-environ==0.10.0
django-tailwind==3.4.0
django-render-partial==0.4
django-allauth==0.54.0
django-birthday==0.1.4
django-phonenumber-field[phonenumbers]==7.1.0
django-money==3.2.0
django-cities-light==3.9.2
django-smart-selects==1.6.0
django-multiselectfield==0.1.12
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-ipware==5.0.0
django-filter==23.2
django-widget-tweaks==1.4.12
django-cleanup==8.0.0
Pillow==10.0.0
file_validator==1.0.9
geoip2==4.7.0
sorl-thumbnail==12.9.0

# Test
pytest==7.4.0
coverage==7.2.7

# Development
pylint
black